senthil/requirements.txt
flask==2.0.1
pymongo==3.12.0
twilio==7.12.0
python-dotenv==0.19.0
opencv-python==4.5.3.56
numpy==1.21.2
inference-sdk==0.9.0